question2/worker-service/requirements.txt
# Core dependencies
psycopg2-binary==2.9.7
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.1.0

# Development dependencies
pytest==7.4.0
pytest-cov==4.1.0
